Remove trailing whitespace
Several lines end up with trailing whitespace if a value is not provided. For instance, if no tags are provided, that line becomes `.TAGS ` (note the trailing whitespace). This violates default PSScriptAnalyzer rules, and is otherwise an inconvenience. This change will append the correct whitespace and the value if it's provided, and prevent extraneous whitespace from being added if no value is provided (as in the field is being left empty/blank).