The scope of all PowerShell-Core scripts/modules was always changed from AllUsers to CurrentUsers.
This is not the expected behaviour of course. if I installed a PowerShell-Core script for all users I'd expect it to still be installed for all users after calling Update-Script, which wasn't the case currently.
So the check for $script:IsCoreCLR was removed from the scope, since this ment for PowerShell-Core scripts/modules that they'd always be reinstalled into the user scope during updates, no matter the original scope of the PowerShell Core script/module.
Also bumped version number so it can be deployed.